requirements.txt
yfinance~=0.2.32
pandas~=2.1.3
functions-framework==3.*
google-cloud-firestore
flask~=3.0.0